+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### History of major changes to mpes since the first pip-installable version
 
+#### 2017.08.31
+1. Now axes handle can be passed to colormesh2d()
+2. Added replist to mpes.utils
+3. Added fit_parameter_plot() to mpes.visualization (built on top of colormesh2d)
+
 #### 2017.08.29
 1. Added colormap normalization